pyproject.toml
[project]
name = "credit-card-fraud-detection"
version = "0.1.0"
description = "A cloud-native, end-to-end real-time fraud detection system leveraging Apache Spark Streaming, Kafka, Cassandra, and Airflow, orchestrated entirely on Kubernetes"
requires-python = ">=3.11"
authors = ["@gagansingh894"]
dependencies = [
    "apache-airflow==2.7.3",
    "apache-airflow-providers-apache-spark>=4.9.0",
    "cassandra-driver<=3.29",
    "confluent-kafka==2.10",
    "mlflow==3.2",
    "numpy==2.2",
    "pydantic==2.10",
    "pyspark<=3.5.6",
]

